introuce AssignExtrinsicTypes trait
