+++ src/extrinsic/balances.rs
@@ -17,48 +17,36 @@
 
 //! Extrinsics for `pallet-balances`.
 
+use super::{AssignExtrinsicTypes, ExtrinsicFor};
 use crate::{api::Api, rpc::Request};
 use ac_compose_macros::compose_extrinsic;
-use ac_primitives::{
-	BalancesConfig, CallIndex, ExtrinsicParams, SignExtrinsic, UncheckedExtrinsicV4,
-};
+use ac_primitives::{BalancesConfig, CallIndex, ExtrinsicParams, SignExtrinsic};
 use alloc::borrow::ToOwned;
 use codec::{Compact, Encode};
-use serde::de::DeserializeOwned;
 
 pub const MODULE: &str = "Balances";
 pub const TRANSFER: &str = "transfer";
 pub const SET_BALANCE: &str = "set_balance";
 
-pub type BalanceTransferCall<Address, Balance> = (CallIndex, Address, Compact<Balance>);
-pub type BalanceSetBalanceCall<Address, Balance> =
+pub type TransferCall<Address, Balance> = (CallIndex, Address, Compact<Balance>);
+pub type SetBalanceCall<Address, Balance> =
 	(CallIndex, Address, Compact<Balance>, Compact<Balance>);
 
-type AddressFor<Module> = <Module as CreateBalancesExtrinsic>::Address;
-type SignatureFor<Module> = <Module as CreateBalancesExtrinsic>::Signature;
-type SignedExtraFor<Module> = <Module as CreateBalancesExtrinsic>::SignedExtra;
-
-type ExtrinsicFor<Module, Call> =
-	UncheckedExtrinsicV4<AddressFor<Module>, Call, SignatureFor<Module>, SignedExtraFor<Module>>;
-
-pub trait CreateBalancesExtrinsic {
+pub trait CreateBalancesExtrinsic: AssignExtrinsicTypes {
 	type Balance;
-	type Address;
-	type Signature;
-	type SignedExtra;
 
 	fn balance_transfer(
 		&self,
 		to: Self::Address,
 		amount: Self::Balance,
-	) -> ExtrinsicFor<Self, BalanceTransferCall<Self::Address, Self::Balance>>;
+	) -> ExtrinsicFor<Self, TransferCall<Self::Address, Self::Balance>>;
 
 	fn balance_set_balance(
 		&self,
 		who: Self::Address,
 		free_balance: Self::Balance,
 		reserved_balance: Self::Balance,
-	) -> ExtrinsicFor<Self, BalanceSetBalanceCall<Self::Address, Self::Balance>>;
+	) -> ExtrinsicFor<Self, SetBalanceCall<Self::Address, Self::Balance>>;
 }
 
 impl<Signer, Client, Params, Runtime> CreateBalancesExtrinsic
@@ -69,18 +57,14 @@ where
 	Runtime: BalancesConfig,
 	Params: ExtrinsicParams<Runtime::Index, Runtime::Hash>,
 	Compact<Runtime::Balance>: Encode,
-	Runtime::Header: DeserializeOwned,
 {
 	type Balance = Runtime::Balance;
-	type Address = Signer::ExtrinsicAddress;
-	type Signature = Signer::Signature;
-	type SignedExtra = Params::SignedExtra;
 
 	fn balance_transfer(
 		&self,
 		to: Self::Address,
 		amount: Self::Balance,
-	) -> ExtrinsicFor<Self, BalanceTransferCall<Self::Address, Self::Balance>> {
+	) -> ExtrinsicFor<Self, TransferCall<Self::Address, Self::Balance>> {
 		compose_extrinsic!(self, MODULE, TRANSFER, to, Compact(amount))
 	}
 
@@ -89,7 +73,7 @@ where
 		who: Self::Address,
 		free_balance: Self::Balance,
 		reserved_balance: Self::Balance,
-	) -> ExtrinsicFor<Self, BalanceSetBalanceCall<Self::Address, Self::Balance>> {
+	) -> ExtrinsicFor<Self, SetBalanceCall<Self::Address, Self::Balance>> {
 		compose_extrinsic!(
 			self,
 			MODULE,
